+++ src/api/editor.ts
@@ -114,7 +114,7 @@ export function PATCH(args: rapi_contracts.ApiMethodArguments): PromiseLike<any>
         try {
             let editor = vscode.window.activeTextEditor;
             if (editor) {
-                rapi_helpers.readHttpBody(args.request.request).then((body) => {
+                args.getBody().then((body) => {
                     try {
                         let str = (body || Buffer.alloc(0)).toString('utf8');